requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
flask_wtf==1.2.1
pyOpenSSL==24.2.1
pyotp==2.9.0
qrcode==7.4.2
Requests==2.32.3
WTForms==3.1.2
email_validator==2.2.0
gunicorn==23.0.0
Werkzeug==3.0.4
flask-sock==0.7.0
websocket-client==1.9.0